requirements.txt
Flask==2.3.3
moviepy==1.0.3
Pillow==9.5.0
python-dotenv==1.0.0
assemblyai